添加全面查询

+++ app/controler/blog.ts
@@ -4,20 +4,17 @@ import { BlogParams } from "../types/index";
 import blogService from "../service/blog";
 import Blog from "../models/blog";
 import _ from "lodash";
+import Utils from "../utils/index";
 import { Op } from "sequelize";
 
 class BlogControler {
   async add(ctx: Context) {
     let { title, category, content, created_at } = ctx.request
       .body as BlogParams;
+    let params = { title, category, content, created_at };
 
     let result = await Blog.findOne({
-      where: {
-        created_at: {
-          [Op.gt]: new Date("2023-03-23"),
-          [Op.lt]: new Date("2023-03-24"),
-        },
-      },
+      where: { ...Utils.formatParams(params) },
     });
 
     if (result) {
@@ -37,19 +34,32 @@ class BlogControler {
 
   async find(ctx: Context) {
     let params = ctx.query;
+    params.user_id = ctx.state.user_id;
 
-    let data = (await Blog.findAll({ where: { ...params } })) as {}[];
+    let struct = {
+      title: "like",
+      category: "eq",
+      content: "like",
+      sort_id: "eq",
+      user_id: "eq",
+      created_at: "range",
+    };
+
+    let addition = Utils.generateSQL(struct, Utils.formatParams(params));
+    if (params.created_at) {
+      let { startTime, endTime } = Utils.getRangeTimeByTimeStamp(
+        params.created_at
+      );
+      addition["created_at"] = { [Op.gt]: startTime, [Op.lt]: endTime };
+    }
+
+    let data = (await Blog.findAll({ where: { ...addition } })) as {}[];
     SUCCESS(ctx, data);
   }
 
   async remove(ctx: Context) {
     let id = ctx.params.id;
-
-    await blogService.remove({
-      user_id: ctx.state.id,
-      id: id,
-    });
-
+    await blogService.remove({ user_id: ctx.state.id, id: id });
     SUCCESS(ctx);
   }
